Remove old regulations work history controller
This removes the old work history controller as it existed under the old
regulations and replaces it with the one built for the new regulations.
It's unlikely we'll ever need the old code as the new regulations have
now launched, but if we do we can always revert this commit.
