--- app/views/shared/application_form/_new_regs_work_history_summary.html.erb
@@ -1,53 +1,0 @@
-<% work_histories.each do |work_history| %>
-  <% application_form = work_history.application_form %>
-  <%= render(CheckYourAnswersSummary::Component.new(
-    id: "work-history-#{work_history.id}",
-    model: work_history,
-    title: work_history.school_name,
-    fields: {
-      school_name: {
-        href: [:school, :teacher_interface, :application_form, :new_regs, work_history]
-      },
-      city: {
-        title: "City of institution",
-        href: [:school, :teacher_interface, :application_form, :new_regs, work_history]
-      },
-      country_name: {
-        title: "Country of institution",
-        href: [:school, :teacher_interface, :application_form, :new_regs, work_history]
-      },
-      job: {
-        title: "Your job role",
-        href: [:school, :teacher_interface, :application_form, :new_regs, work_history]
-      },
-      hours_per_week: {
-        title: "Hours per week",
-        href: [:school, :teacher_interface, :application_form, :new_regs, work_history]
-      },
-      start_date: {
-        title: "Role start date",
-        format: :without_day,
-        href: [:school, :teacher_interface, :application_form, :new_regs, work_history]
-      },
-      end_date: !work_history.still_employed ? {
-        title: "Role end date",
-        format: :without_day,
-        href: [:school, :teacher_interface, :application_form, :new_regs, work_history]
-      } : nil,
-      contact_name: !application_form.reduced_evidence_accepted ? {
-        title: "Contact name",
-        href: [:contact, :teacher_interface, :application_form, :new_regs, work_history]
-      } : nil,
-      contact_job: !application_form.reduced_evidence_accepted ? {
-        title: "Contact job",
-        href: [:contact, :teacher_interface, :application_form, :new_regs, work_history]
-      } : nil,
-      contact_email: !application_form.reduced_evidence_accepted ? {
-        title: "Contact email address",
-        href: [:contact, :teacher_interface, :application_form, :new_regs, work_history]
-      } : nil,
-    },
-    changeable:,
-    delete_link_to: [:delete, :teacher_interface, :application_form, :new_regs, work_history]
-  )) %>
-<% end %>
