--- app/views/teacher_interface/new_regs/work_histories/check_member.html.erb
@@ -1,15 +1,0 @@
-<% content_for :page_title, t(".title") %>
-<% content_for :back_link_url, back_history_path(origin: true, default: teacher_interface_application_form_new_regs_work_histories_path) %>
-
-<span class="govuk-caption-l"><%= t("application_form.tasks.sections.work_history") %></span>
-<h1 class="govuk-heading-l"><%= t(".title") %></h1>
-
-<%= render "shared/application_form/new_regs_work_history_summary",
-           work_histories: [@work_history],
-           changeable: true %>
-
-<% if @next_work_history %>
-  <%= govuk_button_link_to "Continue", [:school, :teacher_interface, :application_form, :new_regs, @next_work_history] %>
-<% else %>
-  <%= govuk_button_link_to "Continue", %i[add_another teacher_interface application_form new_regs work_histories] %>
-<% end %>
