--- app/views/teacher_interface/new_regs/work_histories/new.html.erb
@@ -1,4 +1,0 @@
-<% content_for :page_title, "#{"Error: " if @form.errors.any?}#{t("application_form.work_history.school.heading.#{@work_history.locale_key}")}" %>
-<% content_for :back_link_url, back_history_path(default: teacher_interface_application_form_new_regs_work_histories_path) %>
-
-<%= render "school_form", form: @form %>
