+++ app/views/teacher_interface/work_histories/_contact_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with model: form, url: [:contact, :teacher_interface, :application_form, :new_regs, form.work_history], method: :post do |f| %>
+<%= form_with model: form, url: [:contact, :teacher_interface, :application_form, form.work_history], method: :post do |f| %>
   <%= f.govuk_error_summary %>
 
   <h1 class="govuk-heading-l">
