--- app/views/teacher_interface/work_histories/_heading.html.erb
@@ -1,5 +1,0 @@
-<span class="govuk-caption-l">Proof that you're recognised as a teacher</span>
-<h1 class="govuk-heading-l">Your work history in education</h1>
-
-<p class="govuk-body">We need to see your work history in education, beginning with your current or most recent employer.</p>
-<p class="govuk-body">You can add previous employers and roles dating back to a maximum of 10 years.</p>
