+++ app/views/teacher_interface/work_histories/delete.html.erb
@@ -1,15 +1,14 @@
-<% content_for :page_title, "Delete work history" %>
+<% content_for :page_title, t(".title") %>
 <% content_for :back_link_url, back_history_path(default: teacher_interface_application_form_work_histories_path) %>
 
-<%= form_with model: @delete_work_history_form, url: [:teacher_interface, :application_form, @work_history], method: :delete do |f| %>
+<%= form_with model: @form, url: [:teacher_interface, :application_form, @work_history], method: :delete do |f| %>
   <%= f.govuk_error_summary %>
 
   <%= f.govuk_collection_radio_buttons :confirm,
-                                       [OpenStruct.new(id: :true, name: "Yes"), OpenStruct.new(id: :false, name: "No")],
-                                       :id,
-                                       :name,
+                                       [OpenStruct.new(value: :true), OpenStruct.new(value: :false)],
+                                       :value,
                                        inline: true,
                                        legend: { text: "Are you sure you want to delete #{work_history_title(@work_history)}?", size: "l", tag: "h1" } %>
 
-  <%= f.govuk_submit "Continue", prevent_double_click: false %>
+  <%= f.govuk_submit %>
 <% end %>
