+++ app/views/teacher_interface/work_histories/edit_contact.html.erb
@@ -1,4 +1,4 @@
 <% content_for :page_title, "#{"Error: " if @form.errors.any?}#{t("application_form.work_history.contact.heading")}" %>
-<% content_for :back_link_url, back_history_path(default: teacher_interface_application_form_new_regs_work_histories_path) %>
+<% content_for :back_link_url, back_history_path(default: teacher_interface_application_form_work_histories_path) %>
 
 <%= render "contact_form", form: @form %>
