+++ app/views/teacher_interface/work_histories/requirements_unmet.html.erb
@@ -1,5 +1,5 @@
 <% content_for :page_title, t("application_form.tasks.sections.work_history") %>
-<% content_for :back_link_url, back_history_path(default: teacher_interface_application_form_new_regs_work_histories_path) %>
+<% content_for :back_link_url, back_history_path(default: teacher_interface_application_form_work_histories_path) %>
 
 <h1 class="govuk-heading-l">Your work experience does not meet all of the requirements for QTS</h1>
