+++ spec/support/autoload/page_objects/teacher_interface/check_work_histories.rb
@@ -1,7 +1,7 @@
 module PageObjects
   module TeacherInterface
     class CheckWorkHistories < SitePrism::Page
-      set_url "/teacher/application{/new_regs}/work_histories/check"
+      set_url "/teacher/application/work_histories/check"
 
       element :heading, "h1"
       sections :summary_list_cards, GovukSummaryListCard, ".govuk-summary-card"
