+++ spec/support/autoload/page_objects/teacher_interface/check_work_history.rb
@@ -1,7 +1,7 @@
 module PageObjects
   module TeacherInterface
     class CheckWorkHistory < SitePrism::Page
-      set_url "/teacher/application{/new_regs}/work_histories/{work_history_id}/check"
+      set_url "/teacher/application/work_histories/{work_history_id}/check"
 
       element :heading, "h1"
       section :summary_list, GovukSummaryList, ".govuk-summary-list"
