+++ spec/support/autoload/page_objects/teacher_interface/edit_work_history_school.rb
@@ -3,7 +3,7 @@
 module PageObjects
   module TeacherInterface
     class EditWorkHistorySchool < WorkHistorySchoolForm
-      set_url "/teacher/application/new_regs/work_histories/{work_history_id}/school"
+      set_url "/teacher/application/work_histories/{work_history_id}/school"
     end
   end
 end
