+++ spec/support/autoload/page_objects/teacher_interface/new_work_history.rb
@@ -3,7 +3,7 @@
 module PageObjects
   module TeacherInterface
     class NewWorkHistory < WorkHistorySchoolForm
-      set_url "/teacher/application/new_regs/work_histories/new"
+      set_url "/teacher/application/work_histories/new"
     end
   end
 end
